refactor(core): remove request validation
Removing request validation from the core package. The manual validation was intended
to take stress from the API. However, it takes too much effort to keep up with such
atomic changes in the API. Leaving it to the API to do the validation. This cleans
up the code and makes it more flexible in the future.